Applied AbstractCards -> Cards.py, removed AbstractCards.py; modified methods in Agent, Field, Game to accommodate new Cards.py
