+++ game/Cards.py
@@ -1,183 +1,128 @@
 import numpy as np
-import random
 from collections import deque
+import random
 
-_CARD_MAP_FULL = np.array([['D0', 'D1', 'D2', 'D3', 'D4', 'D5', 'D6', 'D7', 'D8', 'D9', 'D10', 'D11', 'D12'],
-                 ['C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'C8', 'C9', 'C10', 'C11', 'C12'],
-                 ['H0', 'H1', 'H2', 'H3', 'H4', 'H5', 'H6', 'H7', 'H8', 'H9', 'H10', 'H11', 'H12'],
-                 ['S0', 'S1', 'S2', 'S3', 'S4', 'S5', 'S6', 'S7', 'S8', 'S9', 'S10', 'S11', 'S12']])
+_CARD_MAP_FULL = np.array([['D0', 'D1', 'D2', 'D3', 'D4', 'D5', 'D6', 'D7', 'D8', 'D9', 'D  10', 'D11', 'D12'],
+                   ['C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'C8', 'C9', 'C10', 'C11'  , 'C12'],
+                   ['H0', 'H1', 'H2', 'H3', 'H4', 'H5', 'H6', 'H7', 'H8', 'H9', 'H10', 'H11'  , 'H12'],
+                   ['S0', 'S1', 'S2', 'S3', 'S4', 'S5', 'S6', 'S7', 'S8', 'S9', 'S10', 'S11'  , 'S12']])
+_SUITS = ['Diamonds', 'Clubs', 'Hearts','Spades']
+_VALUES = ['2','3','4','5','6','7','8','9','10','Jack','Queen','King','Ace']
 
-_CARD_MAP_SMALL = np.array([['D0', 'D1', 'D2', 'D3', 'D4', 'D5', 'D6', 'D7', 'D8'],
-                 ['C0', 'C1', 'C2', 'C3', 'C4', 'C5', 'C6', 'C7', 'C8'],
-                 ['H0', 'H1', 'H2', 'H3', 'H4', 'H5', 'H6', 'H7', 'H8'],
-                 ['S0', 'S1', 'S2', 'S3', 'S4', 'S5', 'S6', 'S7', 'S8']])
-_SUITS = ['Diamonds', 'Clubs', 'Hearts', 'Spades']
-_VALUES = ['2','3','4','5','6','7','8','9','10','J','Q','K','A']
+class Card():
+    """Map keys: (suit, value) to value (string inside _CARD_MAP_FULL). Additional string method outputs card as 'value' of 'suit' -- e.g., 'Ace' of 'Spades'."""
+    def __init__(self, suit: int, value: int):
+        self.card = _CARD_MAP_FULL[suit,value]
+        self.suit = suit
+        self.value = value
 
-class CardCollection:
-    def __init__(self, n_suits=4, n_vals=9, fill=False):
-        self.order = None
-        if fill:
-            self.cards = np.ones((n_suits, n_vals)).astype(int)
+    def __repr__(self):
+        return self.card
+
+    def __str__(self):
+        return _VALUES[self.value] + ' of ' + _SUITS[self.suit]
+
+class CardCollection():
+    """First In Last Out collection of Card objects, implemented with collection.deque storing deck order, and np.ndarray representing cards inside the deck."""
+    def __init__(self, n_suit: int, n_vals: int, fill=True):
+        assert(n_suit <=4 and n_vals <= 13), 'Deck dimensions out of bounds'
+        if fill == True:
+            self.deck = deque([Card(s,v) for s in range(n_suit) for v in range(n_vals)])
+            self.cards = np.ones((n_suit, n_vals))
         else:
-            self.cards = np.zeros((n_suits, n_vals)).astype(int)
-        self.reorder()
+            self.deck = deque()
+            self.cards = np.zeros((n_suit, n_vals))
+        self.n_suit = n_suit
+        self.n_vals = n_vals
 
-    def __getitem__(self, idx):
-        """Returns the string representation of card at specified position."""
-        if self.__len__() == 0:
-            print("Deck is empty.")
-            return None
-        if self.cards.size == 52:
-            return _CARD_MAP_SMALL[self.order[idx]]
-        if self.cards.size == 36:
-            return _CARD_MAP_FULL[self.order[idx]]
+    def __repr__(self):
+        return repr(self.deck)
 
-    def __eq__(self, other):
-        self.order = other.order
-        self.cards = other.cards
-
-    def __add__(self, other):
-        """This follows the convention that the other CardCollection we're adding is going on top of this CardCollection."""
-        self.cards += other.cards
-        self.order = other.order + self.order
-        other.empty()
-        assert(self.cards.all() == 0 or self.cards.all() == 1), 'Oops! we have 2 or more of the same card.'
-        return self
-
-    # def __iter__(self):
-    #     if self.mode == 'full':
-    #         yield from _CARD_MAP_FULL[self.order]
-    #     else:
-    #         yield from _CARD_MAP_SMALL[self.order]
-
-    # def __next__(self):
-    #     pass
     def __len__(self):
-        """Length of deck."""
-        return len(self.order)
+        return len(self.deck)
 
     def __str__(self):
         head = '--- Card Collection ---\n'
-
-        if self.cards.size == 52:
-            card_str = 'Cards: ' + ','.join([_CARD_MAP_FULL[idx] for idx in self.order]) + '.\n'
-        elif self.cards.size == 36:
-            card_str = 'Cards: ' + ','.join([_CARD_MAP_SMALL[idx] for idx in self.order]) + '.\n'
-        else:
-            ValueError('INVALID CARDS SIZE')
+        card_str = 'Cards: ' + ','.join([repr(x) for x in self.deck]) +   '.\n'
         size_str = 'Size: ' + str(self.__len__()) + '\n'
         tail = '-----------------------\n'
         return head + card_str + size_str + tail
 
-    def suit(self, idx:int) -> str:
-        """Return the suit of the card at specified position.
+    def __getitem__(self, idx: int):
+        """Get the idx'th card of the deck."""
+        return self.deck[idx]
 
-        Args:
-            idx (int): position at which we want to know the suit.
-        Returns:
-            suit (str): the suit of the card at specified position.
-        """
-        if self.__len__() == 0:
-            print("Deck is empty.")
-            return None
-        return _SUITS[self.order[idx][0]]
-
-    def value(self, idx:int) -> int:
-        """Return the value of the card at specified position.
-
-        Args:
-            idx (int): position at which we want to know the card value.
-        Returns:
-            value (int): the value of the card at specified position.
-        """
-        if self.__len__() == 0:
-            print("Deck is empty.")
-            return None
-        return _VALUES[self.order[idx][1]]
-
-    def draw_card(self,n=1):
-        """Draw n cards from the top of the deck."""
-        #drawn_card = np.zeros((self.n_suits, self.n_vals)).astype(int)
-        drawn_card = CardCollection(n_suits = self.cards.shape[0], n_vals = self.cards.shape[1])
-        drawn_card.empty()
-
-        if self.__len__() == 0:
-            return drawn_card
-
-        for i in range(n):
-            idx = self.order.popleft()
-            self.cards[idx] = 0
-            drawn_card.cards[idx] = 1
-            drawn_card.order.appendleft(idx)
-        return drawn_card
-
-    def empty(self):
-        """ Empty this deck. """
-        self.cards *= 0
-        self.order = deque()
+    def __add__(self, other):
+        self.cards += other.cards
+        self.deck = other.deck + self.deck
+        other.empty()
         return self
 
-    def reorder(self):
-        self.order = deque([i for i,v in np.ndenumerate(self.cards) if v == 1])
+    def suit(self, idx: int) -> str:
+        if 0 <= idx < self.__len__():
+            return self.deck[idx].suit
+        else:
+            return 'Position out of range'
 
-    def shuffle(self):
-        """Randomly shuffle the ordering."""
-        random.shuffle(self.order)
+    def value(self, idx: int) -> str:
+        if 0 <= idx < self.__len__():
+            return self.deck[idx].value
+        else:
+            return 'Position out of range'
+
+    def empty(self):
+        """Empty this deck."""
+        self.cards *= 0
+        self.order = deque()
+
+    def draw_card(self, n=1):
+        drawn_cards = CardCollection(n_suits = self.n_suits, n_vals = self.n_vals, fill=False)
+        if self.__len__() == 0:
+            return drawn_cards
+        for i in range(n):
+            card = self.order.popleft()
+            self.cards[card.suit, card.value] = 0
+            drawn_cards[card.suit, card.value] = 1
+            drawn_cards.deck.appendleft(card)
+        return drawn_cards
 
     def cut(self, idx=None):
-        """Split the deck in two, then re-order.
-
-        Args:
-            idx (int): when None, will choose a random place to cut. Otherwise, will cut at specified index.
-        """
         if idx == None:
-            idx = random.choice(range(len(self.order)))
-        self.order.rotate(idx)
+            idx = random.randint(0,len(self.deck)-1)
+        self.deck.rotate(idx)
 
 class DurakDeck(CardCollection):
-    """Create a card collection, in either small or full mode."""
-    def __init__(self, cards = None, mode='small'):
-        self.mode = mode
-        self.n_suits = None
-        self.n_vals = None
-        self._apply_mode()
+    def __init__(self, mode='small',fill=True):
+        if (mode == 'small'):
+            super().__init__(n_suit = 4, n_vals = 9, fill=fill)
+        if (mode == 'full'):
+            super().__init__(n_suit = 4, n_vals = 13, fill=fill)
 
-        if cards is None:
-            super().__init__(n_suits=self.n_suits, n_vals=self.n_vals, fill=True)
+class DurakHand():
+    """Initialize a Durak hand by drawing 6 cards from a given deck. The hand itself is 4 x N, N being number of values in input deck. The 0'th row always corresponds to the trump suit."""
+    def __init__(self, deck: DurakDeck):
+        my_deck = deck.draw_card(6)
+        self.hand = np.zeros_like(my_deck.cards)
+        self.get_hand_from_deck(my_deck)
+        self.trump_idx = deck[-1].suit
+        self.trump_suit = deck.suit(-1)
+
+    def hand_from_deck(self, deck):
+        indices = list(range(deck.n_suits))
+        indices[0], indices[self.trump_idx] = indices[self.trump_idx], indices[0]
+        self.hand += deck.cards[indices]
+
+    def deck_from_hand(self) -> DurakDeck:
+        indices = list(range(self.hand.shape[0]))
+        indices[0], indices[self._trump_idx] = indices[self._trump_idx], indices[0]
+        if self.hand.size == 52:
+            deck = DurakDeck(cards = self.hand[indices], mode = 'full')
+        elif self.hand.size == 36:
+            deck = DurakDeck(cards = self.hand[indices], mode = 'small')
         else:
-            self.cards = cards
-            self.order = None
-            self.reorder()
+            raise ValueError('INVALID HAND SIZE')
+        return deck
 
-
-        if self.__len__() != 0:
-            self.trump_suit = self.suit(-1)
-            self.trump_idx = self.order[-1][0]
-        else:
-            self.trump_suit = None
-            self.trump_idx = None
-
-    def _apply_mode(self):
-        if self.mode == 'full':
-            self.n_suits = 4
-            self.n_vals = 13
-        elif self.mode == 'small':
-            self.n_suits = 4
-            self.n_vals = 9
-
-    def draw_card(self,n=1):
-        """Draw n cards from the top of the deck."""
-        #drawn_card = np.zeros((self.n_suits, self.n_vals)).astype(int)
-        drawn_card = DurakDeck(mode = self.mode)
-        drawn_card.empty()
-
-        if self.__len__() == 0:
-            return drawn_card
-
-        for i in range(min(n, len(self.order))):
-            idx = self.order.popleft()
-            self.cards[idx] = 0
-            drawn_card.cards[idx] = 1
-            drawn_card.order.appendleft(idx)
-        return drawn_card
+class SpadesDeck(CardCollection):
+    def __init__(self, fill=True):
+        super().__init__(n_suit=4, n_vals=13, fill=fill)
